Fix APIDefinitionFromSwagger20Test test case
